Consolidate rendering configuration templates
The (mostly) redundant 'DeviceRenderConfigView' and
'VirtualMachineRenderConfigView' have been merged into a single
'ObjectRenderConfigView'.
