+++ netbox/templates/extras/object_render_config.html
@@ -1,4 +1,5 @@
-{% extends 'dcim/device/base.html' %}
+{% extends base_template %}
+{% load helpers %}
 {% load static %}
 {% load i18n %}
 
@@ -67,7 +68,7 @@
         {% endif %}
       {% else %}
         <div class="alert alert-info">
-          {% trans "No configuration template has been assigned for this device." %}
+          {% trans "No configuration template has been assigned." %}
         </div>
       {% endif %}
     </div>
